Unexpectedly massive rewrite. The purpose of these changes is to make the script support the new architecture lines in the mirror list, and display notes where possible/appropriate in its output.
This also includes everything we should need to support adding amd64 to the
mirror list, including the ability to filter it off of the website until we
decide to turn it on. It should also be most of the way to supporting SCC.

Along the way, I ended up needing to change the %mirror hash into an array,
to deal with the fact that with amd64 added, some mirrors may appear on the
list more than once, with different arches specified, probably in different
places.

To help in changing that hash to an array, I made the whole program use
strict clean.

Also fixes a minor bug in display of sponsors that have no url, turned up
when use strict was enabled.