Added English source to dependencies
This is required as translation directories just include the Makefile from
english/ instead of writing a new one. Nevertheless the translation check
header option maxdelta= requires this to output something as
"Attention! This translation is outdated, please use the original."

See also the mail "Important: maxdelta and mindelta in translation check header
and missing Makefile dependencies" on debian-www.