Added link to Pure Blends section to front page
The Pure Blends section has been under development for a while now. Yesterday,
I seperated out the data in the page from the core of the page and made it
easier for translators to translate. This morning I added a redirect from
http://blends.debian.org/ to the /blends section of the website and now it is
time to add a link from the Debian front page.