Include PayPal in the list of SPI donation methods
Reduces confusion for people who scroll down, click on the PayPal item in the
methods table and find that it jumps back up to the top of the page.

Also clarify some of the language and fix tags.

Adjust the supplied patch to reduce work for translators.

Patch-mostly-by: Louis-Philippe Véronneau <pollito@riseup.net>
Suggested-in: <6e446927-e4d1-c241-35a8-1e9695ade4c2@riseup.net>