Correct tests for Debian releases
Previously the debian_release test was tied to a specific point
release for each release. This incorrectly narrowed the definitions.
Now we only go off the first digit for the test.

This introduces two new issues. First, the previous releases that
have the same major version number are treated the same. Second,
when we move to 10.X we will have to change it again.